engine/requirements.txt
# MIKE-1 Engine Dependencies

# =============================================================================
# BROKER API
# =============================================================================

# Alpaca Markets - Official API for stocks, options, crypto
# Docs: https://docs.alpaca.markets/
alpaca-py>=0.13.0

# =============================================================================
# CONFIGURATION
# =============================================================================
pyyaml>=6.0
python-dotenv>=1.0.0

# =============================================================================
# DATABASE
# =============================================================================
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# =============================================================================
# DATA & ANALYSIS
# =============================================================================
pandas>=2.0.0
numpy>=1.24.0

# =============================================================================
# HTTP & ASYNC
# =============================================================================
httpx>=0.25.0
aiohttp>=3.9.0

# =============================================================================
# SCHEDULING
# =============================================================================
schedule>=1.2.0

# =============================================================================
# LOGGING
# =============================================================================
structlog>=23.0.0

# =============================================================================
# TYPE HINTS
# =============================================================================
pydantic>=2.0.0

# =============================================================================
# TESTING
# =============================================================================
pytest>=7.0.0
pytest-asyncio>=0.21.0

# =============================================================================
# DEV TOOLS
# =============================================================================
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
